chore: rework and add more tests for the `--update` command line option
